:sparkles: 라우터로 검색 페이지 이동 기능 구현
