+++ index.html
@@ -24,7 +24,8 @@
     <!-- sweetAlert2 -->
     <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
     <script src="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.js"></script>
-    <!-- <script type="module" defer src="./js/router.js"></script> -->
+    <script type="module" defer src="./js/router.js"></script>
+    <link rel="stylesheet" href="./scss/product.scss" />
     <title>Document</title>
   </head>
 
@@ -43,7 +44,7 @@
             </ul>
 
             <div class="header-items">
-              <a class="logo" href="./search-result.html">
+              <a class="logo" href="/" data-navigo>
                 <img src="./static/images/logo.png" alt="오뚜기 몰" />
               </a>
               <form class="search">
