+++ js/components/userPage.js
@@ -318,14 +318,12 @@ export const searchPage = /* html */ `
           <h3 class="product-title"></h3>
           <div class="product-wrapper">
             <div class="product-tab">
-              <p class="total-count">총 <span class="count">20</span>건</p>
+              <p class="total-count">총 <span class="count"></span>건</p>
               <ul class="sort-tab">
-                <li class="selected">
-                  <a href="javascript:void(0)">최신순</a>
-                </li>
-                <li><a href="javascript:void(0)">높은 가격순</a></li>
-                <li><a href="javascript:void(0)">낮은 가격순</a></li>
-                <li><a href="javascript:void(0)">할인율순</a></li>
+                <li class="selected"><button>최신순</button></li>
+                <li><button>높은 가격순</button></li>
+                <li><button>낮은 가격순</button></li>
+                <li><button>할인율순</button></li>
               </ul>
             </div>
             <ul class="product-list"></ul>
