+++ search-result.html
@@ -74,7 +74,6 @@
         </div>
       </header>
       <section id="main">
-        <!-- banner swiper -->
         <div class="inner-wrapper">
           <h3 class="product-title"></h3>
           <div class="product-wrapper">
